linode.py: switch to py3 and linode v4 API
All linode-*.py scripts are now a single linode.py. This script takes
command arguments and options. (Although no real options exist... yet.)

The new Linode v4 API seems to be somewhat slower due to the extraneous
API calls due to load-on-demand architecture. The previous job queue was
much faster. The API transition is forced (as of January 2020??) so we
don't have much of a choice.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>